Renamed package to exercise 4
