+++ src/main/java/exercise4/entities/Category.java
@@ -1,4 +1,4 @@
-package exercise3.entities;
+package exercise4.entities;
 
 public enum Category {
