+++ src/main/java/exercise4/service/DiscountDecorator.java
@@ -1,4 +1,4 @@
-package exercise3.service;
+package exercise4.service;
 
 public class DiscountDecorator extends ProductDecorator {
