+++ src/test/java/exercise4/service/ManagingProductTest.java
@@ -1,8 +1,8 @@
-package exercise3.service;
+package exercise4.service;
 
-import exercise3.entities.Category;
-import exercise3.entities.Product;
-import exercise3.repository.InMemoryProductRepository;
+import exercise4.entities.Category;
+import exercise4.entities.Product;
+import exercise4.repository.InMemoryProductRepository;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDate;
